Refactor StimEffectsDamagePatch to use Postfix with manual body part resolution fallback and change buff timing fields from int to float
- Change Prefix to Postfix to allow original method execution before applying custom logic
- Replace direct method delegate with multi-tier resolution: cached method_9 > reflected method_9 > manual fallback
- Implement ManualFindBestBodyPart with full body part candidate checking logic
- Add IsCandidateBodyPart to check for destroyed parts, fractures, and bleeding effects
- Add GetItem

+++ Server/Models/ItemPackDefinition.cs
@@ -41,13 +41,13 @@ public class StimBuff
     public string SkillName { get; set; } = string.Empty;
 
     [JsonPropertyName("value")]
-    public int Value { get; set; }
+    public float Value { get; set; }
 
     [JsonPropertyName("delay")]
-    public int Delay { get; set; } = 1;
+    public float Delay { get; set; } = 1;
 
     [JsonPropertyName("duration")]
-    public int Duration { get; set; } = 300;
+    public float Duration { get; set; } = 300;
 
     [JsonPropertyName("chance")]
     public float Chance { get; set; } = 1;
